requirements.txt
SQLAlchemy~=1.3.23
Werkzeug~=1.0.1
requests~=2.25.1
WTForms~=2.3.3
waitress~=2.0.0
Flask~=1.1.2
flask-login
python-telegram-bot
flask_wtf
SQLAlchemy-serializer
email-validator
